Created data_eda folder
The data_eda folder contains a Jupiter notebook with scatter plots of health outcomes versus the walkability index. There are also sense checks comparing the dummy regressor to the linear regressor for specific health outcomes versus walkability. The conclusion is that it seems to be challenging to obtain a good model.